refactoring(lotto): LottoNumber class 를 제거하고 Lotto class 에서 바로 로또번호를 생성하도록 수정. 랜덤으로 생성하는 방식 수정
